% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{amoniaco_libre_add}
\alias{amoniaco_libre_add}
\title{Agrega el parámetro Amoníaco Libre a tabla alta}
\usage{
amoniaco_libre_add(.data)
}
\arguments{
\item{.data}{Tabla con datos con ciertas características (ej.:
\code{\link{datos_sia}}). Debe incluir los id_parametros 2032, 2018 y 2090
(Temperatura, pH y NH4).}
}
\value{

}
\description{
Agrega el parámetro Amoníaco Libre a tabla alta
}
\examples{
d <- datos_sia \%>\%
  filtrar_datos(id_programa = 10L,
                rango_fechas = 2019,
                id_parametro = c(2032L, 2018L, 2090L),
                tipo_punto_id = 1L) \%>\%
  amoniaco_libre_add()
dplyr::group_by(d, param) \%>\% tsummary(valor)
hist(log10(d[d$id_parametro == 2091L,]$valor))
}
