% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracciones_infambiental.R
\name{ancho}
\alias{ancho}
\title{Formato ancho}
\usage{
ancho(.data)
}
\arguments{
\item{.data}{Tabla de datos obtenida con \code{\link{consulta_muestras}},
posiblemente modificada con valores_numericos y con columnas agregadas por
\code{\link[dplyr]{left_join}}.}
}
\value{

}
\description{
Ensanchar datos provenientes de consulta_muestras. Espera la presencia de
ciertas columnas. Ver detalles.
}
\details{
Espera que existan las columnas 'valor' y 'param'. En caso de no
encontrarlas las creará. La primera contiene los valores de los parámetros
muestreados, ya sea en formato numérico o en texto. La segunda debería
contener el código del parámetro junto a las unidades de medida entre
paréntesis (ej.: \code{'OD (mg/L)'}). En caso de haber una columna llamada
'codigo_nuevo', será utilizada para crear 'param', de lo contrario usará la
columna 'nombre_clave' (de la tabla parametro, de infambientalbd).

Para facilitar la compatibilidad, esta función además verifica la presencia
de otras columnas agregadas en el código de la app iSIA (en el reactive
\code{datos_extraccion}, del server.R): codigo_nuevo, parametro, grupo,
id_tipo_dato y tipo_dato.
}
