% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracciones_infambiental.R
\name{ancho_old}
\alias{ancho_old}
\title{Formato ancho para 2 parámetros}
\usage{
ancho_old(.data, id_x, id_y, par_x = NULL, par_y = NULL, e_sel = TRUE)
}
\arguments{
\item{.data}{data.frame: Datos como vienen del SIA}

\item{id_x}{integer: id_parametro "X". Ejemplo: 2098 (PT)}

\item{id_y}{integer: id_parametro "Y". Ejemplo: 2097 (PO4)}

\item{par_x}{integer: id_parametro "X". Ejemplo: 2098 (PT)}

\item{par_y}{integer: id_parametro "Y". Ejemplo: 2097 (PO4)}

\item{e_sel}{logical: corresponde considerar la columna est_sel}
}
\value{
data.frame ancha.
}
\description{
\strong{OBSOLETA} (aunque se sigue usando de momento)
}
\details{
Aplicar pivot_wider para datos provenientes de consulta_muestras,
pero con ciertas precauciones...
}
