% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{clasif_tipo_dato}
\alias{clasif_tipo_dato}
\title{Asignar categorías a los datos SIA}
\usage{
clasif_tipo_dato(x)
}
\arguments{
\item{x}{Character. En principio puede ser cualquier vector de tipo
character, pero está pensado específicamente para los valores de la columna
\code{valor_minimo_str} de la tabla \code{datos_muestra_parametros} de la
base de datos infambientalbd (SIA).}

\item{metodo}{String. El valor de este argumento define la forma en que se
clasifican los valores de X. Opciones: "simple" o "informe". Ver detalles.}
}
\value{
Lista con dos elementos:

\describe{

\item{valores}{Valores originales "limpios" (puntos en vez de comas para
decimales y sin espacios en blanco al inicio o al final.)}

\item{tipos}{Vector integer (de misma longitud que \code{x}) con los id de
los tipos de datos, tal como se pueden encontrar en la tabla
\code{\link{tipos_de_dato}} (por ahora no está presente en el SIA, sino en
la carpeta \code{sia_apps/data}).}

}
}
\description{
Evalúa vector character que expresan valores numéricos, según las categorías
de la tabla \code{\link{tipos_de_dato}}.
}
\details{
Números con comas (en vez de puntos) como indicador de decimales,
son considerados numéricos.

El reconocimiento de valores equivalentes a "<LD" y "<LC" se basa en los
datos encontrados en la base infambientalbd, por lo que contempla casos
como "LOQ", "LOD" o "ND" (Limit of Quantification, Limit of Detection y No
Detectado, respectivamente).
}
\examples{
x <- c(" 32,87 ", "2.14", "5e-3", "<ld", ">LC", "ld<x<LC", " < 10", "L.O.Q",
       "ND", "L.O, D ", "LC>X>LD")
clasif_tipo_dato(x) \%>\% as.data.frame()
}
