% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codigos_param}
\alias{codigos_param}
\title{Tabla con códigos alternativos para parámetros}
\format{
Tabla con 255 filas y 7 columnas:

\describe{

\item{grupo}{character. Grupo al que pertenece el parámetro}

\item{parametro}{character. Nombre (largo) del parámetro}

\item{codigo_anterior}{character. Nombre (largo) del parámetro, encontrado
anteriormente}

\item{codigo_nuevo}{character. Abreviación nueva propuesta para el
parámetro}

\item{codigo_nuevo_jm}{character. Abreviación nueva propuesta para el
parámetro, sugerencias de Juan M. Barreneche}

\item{obs}{character. Observaciones}

\item{id_parametro}{integer. Identificador único de cada parámetro.}

}
}
\source{
\code{infambientalbd}
}
\usage{
codigos_param
}
\description{
Presenta códigos nuevos para parámetros, creados por Elena Rodó y Amelia
Fabre.
}
\details{
Última modificación: 2020-09-03
}
\keyword{datasets}
