% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles.R
\name{colapsar_secuencia}
\alias{colapsar_secuencia}
\title{Preparar lista para imprimir}
\usage{
colapsar_secuencia(x, conector = " y ", comillas = FALSE)
}
\arguments{
\item{x}{Vector atomico para imprimir.}

\item{conector}{Caracter para conectar el ultimo elemento de la lista
(tipicamente 'y', '&', 'and', etc...)}

\item{comillas}{Determina si se deben agregar comillas a los elementos de
\code{x}}
}
\value{
Vector character. Ver ejemplos.
}
\description{
Crea un string con la secuencia delementos en x separados por comas, con
excepción del último elemento que es separado con un conector ("y", por
defecto).
}
\examples{
cat("Numeros:", colapsar_secuencia(4:8), "\n")
cat("Numeros:", colapsar_secuencia(4:8, ", "), "\n")
cat("Numbers:", colapsar_secuencia(4:8, " & "), "\n")
cat("Numeros:", colapsar_secuencia(4:8, comillas = TRUE), "\n")
}
