% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles.R
\name{con_sia}
\alias{con_sia}
\title{Conexi贸n con SIA}
\usage{
con_sia()
}
\value{
Una conexi贸n tipo DBI de PosgtgreSQL
}
\description{
Genera una conexi贸n con la base de datos infambiental, usando el usuario y
clave de JMB. S贸lo funciona en algunas PCs (ej.: servidor o PC de JMB).
}
\examples{
\dontrun{
con <- con_sia()
sia_parametro <- dplyr::tbl(con, "parametro")
}
}
