% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cuencas_informes}
\alias{cuencas_informes}
\title{Cuencas y subcuencas en informes}
\format{
Tabla con 51 filas y
8 columnas:

\describe{

\item{nombre_subcuenca_informes}{}

\item{codigo_pto}{}

\item{codigo_pto_mod}{}

\item{id_estacion}{}

\item{id_cuenca}{}

\item{cue_nombre}{}

\item{id_programa}{}

\item{nombre_programa}{}

}
}
\usage{
cuencas_informes
}
\description{
Cuencas y subcuencas en informes
}
\seealso{
\code{\link{sia_estacion}}
}
\keyword{datasets}
