% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles.R
\name{demo_lm}
\alias{demo_lm}
\title{Demostración de informe para Laguna Merín}
\usage{
demo_lm(extension = c("html", "pdf", "doc"))
}
\arguments{
\item{extension}{character. Determina el formato de salida del informe: html,
pdf o doc.}
}
\value{
Genera un archivo Rmd con un ejemplo de informe para Laguna Merín
}
\description{
Demostración de informe para Laguna Merín
}
\examples{
demo_lm()
demo_lm('pdf')
demo_lm('doc')
}
