% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internas.R
\name{det_nonum}
\alias{det_nonum}
\title{Determinar cuáles son no numéricos}
\usage{
det_nonum(v)
}
\arguments{
\item{v}{Character con valores potencialmente convertibles a numéricos. No
debe haber espacios en blanco en ninguno de los valores.}
}
\value{

}
\description{
Determina si un valor es no numérico. Excluye valores escritos con notación
científica.
}
\examples{
v <- c("1.347e4", "<LC", "<78", "14.447", "7", "1,3E+02")
siabox:::det_nonum(v)
}
