% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{eti}
\alias{eti}
\title{Crear etiquetas para gráficos}
\usage{
eti(id_parametro, t_eti)
}
\arguments{
\item{id_parametro}{Escalar (integer). Número de id de parámetro (i.e.:
\code{id_parametro}).}

\item{t_eti}{data.frame. Tabla generada con \code{\link{t_eti_add}}}
}
\value{

}
\description{
Devuelve la etiqueta para el parámetro seleccionado por su \code{id}, basado
en la tabla \code{t_eti} (ver detalles).
}
\details{
Si no se provee una tabla \code{t_eti}, la función usará \code{t_eti_add()}
como tabla de referencia.
}
\examples{
\dontrun{
e <- eti(2101, tibble::tibble(
  id_parametro = 2101,
  etiqueta = expression('NO'[2]*' (mg NO'[2] * '-N/L)')
))
ggplot() + geom_blank() + xlab(eti(2101))
ggplot() + geom_blank() + xlab(e)
}
}
\seealso{
\code{\link{t_eti_add}}
}
