% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracciones_infambiental.R
\name{filtrar_datos}
\alias{filtrar_datos}
\title{Filtrar datos_sia}
\usage{
filtrar_datos(
  .data,
  id_programa = 4L,
  id_matriz = 6L,
  rango_fechas = c("2019-01-10", "2019-12-31"),
  id_parametro = c(2032, 2098, 2101, 2017, 2018, 2009, 2000, 2198),
  id_estacion = NULL,
  niveles_est = NULL
)
}
\arguments{
\item{.data}{Tabla con datos extraidos del SIA (\code{\link{datos_sia}} en
principio)}

\item{niveles_est}{}
}
\value{

}
\description{
Función que filtra la tabla \code{\link{datos_sia}} (u otra tabla con un
conjunto de columnas adecuado) de manera similar a la forma en que se extraen
los datos en la aplicación shiny iSIA.
}
\examples{
filtrar_datos(datos_sia, 5, 6)
filtrar_datos(datos_sia, 5, 11)
}
