% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{g_iet_pto}
\alias{g_iet_pto}
\title{Gráfico de IET}
\usage{
g_iet_pto(.data)
}
\arguments{
\item{.data}{Tabla de datos con columnas codigo_pto (nombres de estaciones) e
IET (índice de estado trófico)}
}
\value{

}
\description{
Gráfico de IET
}
\examples{
datos_sia \%>\%
  filtrar_datos(id_programa = 3L, rango_fechas = 2019,
                tipo_punto_id = 1L, id_parametro = 2098) \%>\%
  dplyr::mutate(IET = iet(valor)) \%>\%
  g_iet_pto
}
