% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{d_lon}
\alias{d_lon}
\alias{g_lon_pto}
\alias{g_lon_pto_files}
\title{Graficar valores anuales longitudinales}
\usage{
d_lon(
  .data,
  id_parametro,
  anio,
  ventana_anios = 5L,
  abr_meses = c("Ene", "Feb", "Mar", "Abr", "May", "Jun", "Jul", "Ago", "Set", "Oct",
    "Nov", "Dic")
)

g_lon_pto(
  .data,
  id_parametro,
  anio,
  ventana_anios = 5L,
  colores_meses = (scales::hue_pal())(12),
  horiz
)

g_lon_pto_files(
  .data,
  anio,
  ventana_anios = 5L,
  tabla_horiz,
  path,
  dims = list(ancho = 80, alto = 80, escala = 1.2)
)
}
\arguments{
\item{.data}{data.frame. Datos del SIA (ver \code{\link{datos_sia}} y
ejemplos)}

\item{id_parametro}{integer. Id de uno/escalar (\code{g_mes_pto}) o varios
elementos (\code{g_mes_pto_all}). Se debe(n) corresponder con los ids
encontrados columna homónima de \code{\link{sia_parametro}}}

\item{anio}{intger. Año}

\item{ventana_anios}{integer. Cantidad de años anteriores al anio en
cuestión. Es un lustro por defecto.}

\item{abr_meses}{character. Abreviaciones a ser usadas para los meses en el
eje x de las graficas.}

\item{colores_meses}{character. Paleta de colores para usar en los distintos
meses.}

\item{horiz}{numeric. Vector con dos valores: min y max, para ser graficados
con líneas horizontales. Típicamente representa mínimos y/o máximos
establecidos por decretos.}

\item{tabla_horiz}{data.frame. Debe tener las columnas \code{id_parametro},
\code{valor}, y \code{extremo} tal como en la tabla \code{\link{decreto}}}

\item{path}{character. Ruta de directorio en donde guardar las imágenes
generadas por \code{g_lon_pto}}

\item{dims}{list. Lista con tres elementos: \code{ancho}, \code{alto} y \code{escala}, que
determinan los valores de \code{width}, \code{height} y \code{scale} (los dos primeros, en
milímetros) en el uso de la función \code{\link[ggplot2]{ggsave}} por parte
de \code{g_lon_pto_files_loop}.}
}
\value{

}
\description{
Grafica los valores de los parámetros por estaciones, incluyendo comparación
con años anteriores.
}
\details{
Las estaciones deben estar previamente ordenadas (como factores, no
necesariamente factores ordenados). Específicamente, el campo de los datos
que debe ser un factor es el de \code{codigo_pto}. El argumento \code{orden_est}
de la función \code{\link{filtrar_datos}} puede ayudar a preparar los datos
de esta manera.

El gráfico creado con \code{g_lon} es considerablemente complejo, ya que usa
una leyenda con categorías cualitativamente diferentes (meses en el año de
interés, promedios del año anterior, promedios año de interés y promedios del
lustro previo). Por esta razón, lo que hace internamente la función es
preparar la una data.frame con los datos agrupados y categorizados
correctamente, usando \code{d_lon}, y luego pasa a crear el gráfico con
comandos de \code{ggplot2}.

La función \code{d_lon} puede ser usada independientemente.
}
\section{Functions}{
\itemize{
\item \code{d_lon}: Funcion que (internamente) prepara los datos para
\code{g_lon_pto}

\item \code{g_lon_pto}: Grafica valores anuales longitudinales

\item \code{g_lon_pto_files}: Guarda gráficos de \code{g_lon_pto} en archivos separados
por parámetro.
}}

\examples{
d <- filtrar_datos(datos_sia, id_programa = 5,
                   rango_fechas = c("2012-01-01", "2019-12-31"),
                   id_parametro = c(2032, 2009, 2017, 2018, 2097:2098, 2105),
                   tipo_punto_id = 1)
h <- dplyr::filter(decreto, clase == "1", !is.na(valor))
g_lon_pto(d, 2032, 2019L, horiz = c(min = 5, max = 28))
g_lon_pto(d, 2032, 2019L, horiz = c(min = 5, max = 28), ventana_anios = 7)
g_lon_pto_files(d, 2019L, tabla_horiz = h)
}
