% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{g_lon_pto_files_loop}
\alias{g_lon_pto_files_loop}
\title{Helper para \code{\link{g_lon_pto_files}}}
\usage{
g_lon_pto_files_loop(
  .data,
  anio,
  ventana_anios,
  tabla_horiz,
  directorio,
  pbar = FALSE
)
}
\arguments{
\item{.data}{data.frame. Datos del SIA (ver \code{\link{datos_sia}} y
ejemplos)}

\item{anio}{intger. Año}

\item{ventana_anios}{integer. Cantidad de años anteriores al anio en
cuestión. Es un lustro por defecto.}

\item{tabla_horiz}{data.frame. Debe tener las columnas \code{id_parametro},
\code{valor}, y \code{extremo} tal como en la tabla \code{\link{decreto}}}

\item{directorio}{character. Ruta al directorio donde se guardarán las
imágenes}

\item{pbar}{logical. Define si se usa la capacidad de shiny de mostrar una
barra de progreso}
}
\value{

}
\description{
Helper para \code{\link{g_lon_pto_files}}
}
