% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{g_long_files}
\alias{g_long_files}
\title{Hacer archivos con gráficos g_long}
\usage{
g_long_files(.data, anio, tabla_horiz, path)
}
\value{

}
\description{
Hacer archivos con gráficos g_long
}
\examples{
d <- readRDS("tmp/datos_cuareim.rds") \%>\%
  dplyr::filter(id_parametro \%in\% c(2035, 2017:2018, 2091, 2098, 2111))
h <- dplyr::filter(decreto, clase == "1", !is.na(valor))
g_long_files(d, 2019L, tabla_horiz = h, path = "tmp")
}
