% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{g_mes}
\alias{g_mes}
\title{Graficar un parámetro por mes}
\usage{
g_mes(
  .data,
  id_param = NULL,
  pos_leyenda = "none",
  nombre_clave_param,
  ylab = NULL
)
}
\arguments{
\item{.data}{Tabla con datos}

\item{id_param}{Escalar (integer). Anula nombre_clave_param.}

\item{nombre_clave_param}{Nombre clave del parámetro, tal como aparece en la
tabla parametro de infambientalbd}

\item{ylab}{Etiqueta para el eje y del gráfico resultante. Si este argumento
no es especificado, se utilizará la función \code{eti} para crear la
etiqueta.}
}
\value{

}
\description{
Graficar un parámetro por mes
}
\details{
Requiere de la tabla decreto_long para agregar las líneas
horizontales correspondientes a los valores límites aceptados para la clase
1 de aguas (decreto 253/79).
}
\examples{
dfo <- data.frame(mes = 3:6, codigo_pto = 100200:100203,
                  id_parametro = 2009, valor = c(140, 120, 198, 166))
decreto_long <- readRDS(file.path(ruta_base, "data/decreto_long.rds"))
g_mes(dfo, id_param = 2009, ylab = 'Cond (μS/cm)')
g_mes(dfo, nombre_clave_param = 'Conduc', ylab = 'Cond (μS/cm)')
}
