% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{g_par_esp}
\alias{g_par_esp}
\title{Graficar valores anuales por estacion}
\usage{
g_par_esp(
  .data,
  id_parametro,
  anio,
  colores_meses = (scales::hue_pal())(12),
  horiz = NULL
)
}
\arguments{
\item{.data}{Daos tomados}

\item{colores_meses}{}
}
\value{

}
\description{
Graficar valores anuales por estacion
}
\examples{
d <- readRDS("tmp/datos_cuareim.rds")
decreto_long <- readRDS("../data/decreto_long.rds")
d$mes <- as.integer(d$mes)
d$anio <- as.integer(d$anio)
g_par_esp(d, 2032, 2019L, horiz = c(min = 5, max = 28))
}
