% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{g_par_esp_files}
\alias{g_par_esp_files}
\title{Hacer archivos con gráficos g_par_esp}
\usage{
g_par_esp_files(.data, anio, tabla_horiz = NULL, path = NULL)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Hacer archivos con gráficos g_par_esp
}
\examples{
d <- readRDS("tmp/datos_cuareim.rds") \%>\%
  dplyr::filter(id_parametro \%in\% c(2035, 2017:2018, 2091, 2098, 2111))
h <- dplyr::filter(decreto_long, clase == "1", !is.na(valor))
g_par_esp_files(d, 2019L, tabla_horiz = h, path = "tmp")
}
