% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{iet}
\alias{iet}
\title{Índice de Estado Trófico}
\usage{
iet(x)
}
\arguments{
\item{x}{numeric: valores de concentración de fósforo total, en microgramos
por litro.}
}
\value{
Un vector numérico con los valores del IET.
}
\description{
Índice de Estado Trófico
}
\details{
Cálculo:

\deqn{IET = 10 \times (6-\frac{0.42-0.36 \times ln(\overline{PT 
  (\mu g/L)})}{ln(2)})-20}
}
\examples{
iet(25)
iet(rlnorm(10, meanlog = 3, 1))
}
