% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{largo.default}
\alias{largo.default}
\title{Formato largo}
\usage{
\method{largo}{default}(.data)
}
\arguments{
\item{.data}{data.frame. Formato similar al creado con
\code{link[siabox]{ancho}} a partir de \code{link[siabox]{datos_sia}} o
afines}
}
\value{
data.frame en formato largo.
}
\description{
Pensada para ser la contraparte de \code{link[siabox]{ancho}}.
}
\examples{
d <- filtrar_datos(datos_sia,
                   id_programa = 10L,
                   rango_fechas = c("2019-01-01", "2020-12-31"),
                   id_parametro = c(2099, 2098)) \%>\%
  dplyr::select(codigo_pto, fecha_muestra, id_parametro, param, id_unidad,
                uni_nombre, limite_deteccion, limite_cuantificacion, valor)
da <- ancho(d)
(dal <- largo(da))

dim(d)
dim(dal)

datos_sia \%>\%
  dplyr::filter(id_programa == 4, id_parametro \%in\% c(2017, 2021)) \%>\%
  ancho \%>\%
  largo
}
