% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{largo.planilla}
\alias{largo.planilla}
\title{Planilla a formato largo}
\usage{
\method{largo}{planilla}(.data, tcols = NULL)
}
\arguments{
\item{.data}{data.frame con los datos en formato ancho (planilla SILAD o
template de campo de vSIA)}

\item{tcols}{data.frame opcional: la tabla de columnas creada durante el
proceso de vSIA.}
}
\value{
Devuelve los mismos datos pero en formato largo, con la columna
\code{nfila} agregada, valores (character) en la columna \code{valor_minimo_str}
(nombre usado para que coincida con \code{datos_muestras_parametros} de
infambietnalbd). El parámetro se indica en la columna \code{nombre_clave},
cuando \code{tcols} es null, o \code{id_parametro}, en caso contrario.

Debe tenerse en cuenta que se eliminan las entradas sin valores (i.e.:
\code{NA}). En caso de que en los datos incluyan LD o LC, las entradas elminadas
son aquellas que no tengan valores en \code{valor_minimo_str}, ni en
\code{limite_deteccion}, ni en \code{limite_cuantificacion}
}
\description{
Planilla refiere a planilla con formato SILAD o al template para datos de
campo del vSIA.
}
\details{
Asume que siempre que hay LD también hay valores de LC.

Importante: si \code{tcols} es \code{NULL}, se asume que las columnas parametro, LD y
LC (si es que los datos incluyen límites de detección o cuantificación),
vienen siempre en ese orden, para todos los parámetros y que todos los
parámetros tienen nombres únicos que los distinguen (cosa que tengo patente
que no es necesariamente cierto). O sea que si están los parámetros AlcT y
STF, entonces las columnas estarán en el orden: AlcT (mg CaCO3/L), AlcT LD,
AlcT LC, STF (mg/L), STF LD, STF LC (aunque sí podrían estar todas las
columnas STF antes que las de AlcT; lo que importa es que siempre venga LC
luego de LD y este luego del parámetro en sí.).

Por esta razón, en vSIA, al hacer largo con el paquete de datos (PD), el cual
puede incluir parámetros de campo sin LD o LC, al mismo tiempo que parámetros
de laboratorio que sí incluye esos límites, es necesario que \code{tcols} no sea
\code{NULL}.
}
\examples{
lista_campo <- readRDS("tmp/campo.rds")
largo(lista_campo$datos)
largo(lista_campo$datos, lista_campo$columnas)
lista_lab <- readRDS("tmp/lab.rds")
largo(lista_lab$datos)
largo(lista_lab$datos, lista_lab$columnas$tabla_columnas)
lista_pd <- readRDS("~/R/sia_apps/vSIA/tmp/lista_pd.rds")
largo.planilla(lista_pd$datos, lista_pd$ppd)
datos_sin_ppd <-
  lista_pd$datos[- lista_pd$ppd$ncol[lista_pd$ppd$tabla == "Campo"]]
largo.planilla(datos_sin_ppd)
}
