% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{largo}
\alias{largo}
\title{Pasar a formato ancho}
\usage{
largo(.data, ...)
}
\arguments{
\item{.data}{data.frame. Tabla de datos, con formatos específicos según su
clase.}
}
\value{
Devuelve los mismos datos pero en formato largo: los nombres del
parámetro van en una columna y en otras columnas: valor, LD y LC. Esto
puede variar según los distintos métodos.
}
\description{
Los valores vienen en character cuando se importan tablas de campo o
laboratorio (clase "planilla"), en el contexto de la app vSIA, y son
equivalentes a la columna valor_minimo_str de los datos provenientes del SIA.
De hecho, los valores así como están (a menos que tengan errores y sean
modificados por un usuarie antes de ser ingresados), van directo a esa misma
columna de la tabla datos_muestra_parametros de infambientalbd.
}
\details{
En cambio, cuando se trata de datos de origen "sia", se asume que están en el
formato de \code{\link[siabox]{datos_sia}} (previamente transformados con la
función \code{\link[siabox]{ancho}}), de forma que se espera que los valores
sean numéricos.
}
