% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles.R
\name{limpia_num}
\alias{limpia_num}
\title{Limpiar texto con valores numéricos}
\usage{
limpia_num(x)
}
\arguments{
\item{x}{character. Valores numéricos expresados en varias formas posibles.}
}
\value{
Vector character sin comas o puntos repetidos, simplemente 1 punto,
y sin espacios en blanco.
}
\description{
Quita espacios en blanco, y sustituye comas, comas
repetidas y puntos repetidos por un único punto (indicador de decimales). Se
basa en expresiones regulares y el paquete \code{stringr}.
}
\details{
Combina muy bien con \code{link{extrae_num}}
}
\examples{
limpia_num("2.3")
limpia_num("2..3")
limpia_num("2,3")
limpia_num("2,,,,3")
limpia_num("   2,,,, 3   ")
limpia_num("< 0,015 mg /L")
}
\seealso{
\code{link{extrae_num}}
}
