% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles.R
\name{make_t_eti_base}
\alias{make_t_eti_base}
\title{Tabla de etiquetas base}
\usage{
make_t_eti_base()
}
\value{
Una tibble (\code{tbl_df}) con dos columnas: \code{id_parametro} y
\code{etiqueta}
}
\description{
Crea una tabla de base para hacer etiquetas para gráficos u otros usos,
basándose en las tablas del SIA (infambientalbd).
}
\examples{
make_t_eti_base()
}
