% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{media_geom}
\alias{media_geom}
\title{Media geométrica}
\usage{
media_geom(x)
}
\arguments{
\item{x}{numeric}
}
\value{
Valor numérico equivalente a la media geométrica de x
}
\description{
Media geométrica
}
\examples{
media_geom(c(3, -2.34, 5427))
media_geom(rnorm(6))
media_geom(rnorm(5))
}
