% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracciones_infambiental.R
\name{par_id}
\alias{par_id}
\alias{pro_id}
\alias{est_id}
\alias{mat_id}
\alias{uni_id}
\alias{ins_id}
\alias{dep_id}
\title{Buscadores de id}
\usage{
par_id(patron, ...)

pro_id(patron, ...)

est_id(patron, ...)

mat_id(patron, ...)

uni_id(patron, ...)

ins_id(patron, ...)

dep_id(patron, ...)
}
\arguments{
\item{patron}{character. Patrón (expresión regular tipo
\code{\link[base]{regex}}).}

\item{...}{Argumentos pasados a \code{\link[base]{agrepl}} para buscar en las
columnas \code{parametro} y \code{nombre_clave} de \code{\link{sia_parametro}}}
}
\value{

}
\description{
Buscadores de id para las varias tablas importadas del SIA, usando un texto
(un google de id_parametros). El texto o patrón puede ser una expresión
regular (la cual será evaluada por \code{\link[base]{agrepl}}).
}
\section{Functions}{
\itemize{
\item \code{pro_id}: Busca programas en \code{\link{sia_programa}} en base al
campo \code{nombre_programa} de dicha tabla.

\item \code{est_id}: Busca estaciones en \code{\link{sia_estacion}} en base a
los campos \code{codigo_pto} y \code{estacion} de dicha tabla.

\item \code{mat_id}: Busca matrices en \code{\link{sia_matriz}} en base al
campo \code{nombre} de dicha tabla.

\item \code{uni_id}: Busca unidades en \code{\link{sia_unidad}} en base al
campo \code{uni_nombre} de dicha tabla.

\item \code{ins_id}: Busca instituciones en \code{\link{sia_institucion}} en
base al campo \code{nombre} de dicha tabla.

\item \code{dep_id}: Busca departamentos en \code{\link{sia_departamento}} en
base al campo \code{dep_nombre} de dicha tabla.
}}

\examples{
par_id("fósforo")
par_id("pt", max.distance = 0)
pro_id("merin")
est_id("pascual")
dep_id("cane")
est_id("pascual")
est_id("pascal")
est_id("pscal")
est_id("pacl")
mat_id("agua")
uni_id("mg/l")
uni_id("mg/l", max.distance = 0)
ins_id("cane")
dep_id("flor")
}
\seealso{
\code{\link{sia_parametro}}
}
