% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internas.R
\name{parentesis}
\alias{parentesis}
\title{Auxiliar para \code{consulta_muestras}}
\usage{
parentesis(v, ini = "(", fin = ")", comillas = FALSE)
}
\arguments{
\item{v}{Vector con elementos a concatenar dentro del paréntesis.}

\item{ini}{Inicio del paréntesis.}

\item{fin}{Fin del paréntesis.}

\item{comillas}{logical (1 valor): si poner o no comillas (simples) alrededor
de cada elemento de \code{v}}
}
\value{

}
\description{
Crea un texto con la secuencia de elementos entre paréntesis y separados por
comas.
}
\examples{
siabox:::parentesis(3147:3152)
siabox:::parentesis(3147:3152, "-<<<- |", "| ->>>-")
siabox:::parentesis(3147:3152, comillas = TRUE)
siabox:::parentesis(c("gato", "perro"))
siabox:::parentesis(c("gato", "perro"), comillas = TRUE)
}
