% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internas.R
\name{pegar_obs}
\alias{pegar_obs}
\title{Pegar observaciones}
\usage{
pegar_obs(obs_a, obs_b)
}
\arguments{
\item{obs_a}{character. Vector de observaciones.}

\item{obs_b}{character. Vector de observaciones.}
}
\value{
Vector character con observaciones de uno u otro vector, o, en caso
de que ambos tengan, observacones pegadas.
}
\description{
Pensada para trabajar internamente en \code{link{consulta_muestras}} y en la
creación de \code{link{datos_sia}} (en código de carpeta data-raw)
}
\examples{
a <- c("", NA, "Nubes", "Sin calibrar", "")
b <- c("Lluvia", NA, NA, "Con cuchillo", "")
data.frame(a, b, pegar_obs(a, b))
}
