% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{programa_matriz}
\alias{programa_matriz}
\title{Matrices de programas}
\format{
Tabla con 30 filas y 2
columnas:

\describe{

\item{id_parametro}{integer. Identificador único de cada programa de
monitoreo.}

\item{id_matriz}{Número que indica el id de la matriz asociada al programa
de monitoreo (de momento: 6 = Aguas superficiales y 11 = Sedimentos;
15/10/2020)}

}
}
\source{
\code{infambientalbd}
}
\usage{
programa_matriz
}
\description{
Creada a partir de la tabla \code{\link{sia_estacion}}. Sirve de referencia
para asociar programas de monitoreo con matrices ambientales.
}
\examples{
# Crear una versión actualizada de la tabla:
programa_matriz <-
  sia_estacion \%>\%
    dplyr::distinct(prog_monitoreo, matriz_estacion) \%>\%
    setNames(c("id_programa", "id_matriz")) \%>\%
    dplyr::filter_all(~ !is.na(.))
}
\seealso{
\code{\link{sia_estacion}}
}
\keyword{datasets}
