% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{raiz}
\alias{raiz}
\title{Raíz enésima}
\usage{
raiz(x, n)
}
\arguments{
\item{x}{numeric.}

\item{n}{integer, valor único.}
}
\value{
Devuelve la raíz enésima de un valor \code{x}, incluso cuando ese
valor es negativo, si es que \code{n} es impar. En caso de que \code{n} sea
par y \code{x} negativo, devueve \code{NaN}. Ver ejemplo para comparar con
\code{x ^ (1 / n)} y con \code{abs(x) ^ (1 / n)}.
}
\description{
Raíz enésima
}
\examples{
require(dplyr)
require(tibble)
require(magrittr)
tibble(x = c(4, -4, 4, -4), n = c(2, 2, 3, 3)) \%>\%
  mutate(raiz(x, n), x ^ (1 / n), abs(x) ^ (1 / n))
}
