% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internas.R
\name{set_utf8}
\alias{set_utf8}
\title{Codificar columnas con UTF-8}
\usage{
set_utf8(x)
}
\arguments{
\item{x}{data.frame con columnas character para codificar como UTF-8}
}
\value{
data.frame con los mismos datos pero con columnas character
codificadas con UTF-8
}
\description{
Tomada de esta respuesta en SO: https://stackoverflow.com/a/41900818/1030523.
Notar que tal vez sea buena idea usar RPostgres en vez de RPostgreSQL para
evitar el problema original.
}
