% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_cuenca}
\alias{sia_cuenca}
\title{Lista de cuencas}
\format{
Tabla con 8 filas y
2)` columnas:

\describe{

\item{id}{integer. Número asignado a la cuenca}

\item{cue_nombre}{character. Nombre de la cuenca}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_cuenca
}
\description{
La tabla de cuencas tiene 8 cuencas definidas, incluyendo
categorías extra (Falta Cuenca y Dato Migrado).
}
\details{
Fecha de extraccion: 2021-08-05
}
\examples{
dplyr::left_join(dplyr::select(sia_sub_cuenca, id:sub_cue_cuenca_id),
                 sia_cuenca, by = c('sub_cue_cuenca_id' = 'id'))

}
\keyword{datasets}
