% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_datos_muestra_parametros}
\alias{sia_datos_muestra_parametros}
\title{Tabla con datos de infambientalbd}
\format{
Tabla con 237957 filas y 17 columnas:

\describe{

\item{id_muestra}{integer. Identificador de la muestra (ver sia_muestra)}

\item{id_parametro}{integer. Identificador del parámetro.}

\item{observacion}{character. Observaciones.}

\item{valor_minimo}{numeric. ..completar..}

\item{valor_maximo}{numeric. ..completar..}

\item{tipo_medicion}{integer ..completar..}

\item{id}{integer. Identificador único de cada entrada de la tabla.}

\item{id_estado}{integer. identificador que puede tomar 3 valores: 1.
pendiente, 2. original, 3. aprobado}

\item{fecha_resolucion}{datetime. ..completar..}

\item{valor_minimo_str}{character. Valor ingresado para la muestra del
parámetro. No necesariamente es una expresión numérica. Ver detalles.}

\item{id_estacion}{integer. Identificador de la estación de monitoreo.}

\item{replica}{logical. ..completar..}

\item{nro_muestra}{integer. Número de muestra asignado por SILAD. No es
único para esta tabla (no confundir con id_muestra).}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_datos_muestra_parametros
}
\description{
Esta tabla guarda los datos de cada muestra. Cada fila representa una
combinación única de parámetro, fecha, hora y sitio (estación).
}
\details{
Fecha de extraccion: 2020-10-29.
}
\section{Atencion}{
 La tabla puede tener datos repetidos. La función
\code{\link{consulta_muestras}} se encarga de descartar repeticiones.
}

\section{Valores}{
 Los valores almacenados en la tabla no necesariamente
expresan valores numéricos. Las funciones \code{\link{clasif_tipo_dato}} y
}

\seealso{
\code{\link{consulta_muestras}}
}
\keyword{datasets}
