% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_departamento}
\alias{sia_departamento}
\title{Lista de departamentos}
\format{
Tabla con 21 filas y
3)` columnas:

\describe{

\item{id}{}

\item{dep_nombre}{Nombre del departamento}

\item{dep_codigo}{Código o abreviación del nombre del departamento}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_departamento
}
\description{
La tabla de departamentos tiene los 19 departamentos de Uruguay y 2 entradas
extra: "DATO MIGRADO" y "---" (codigos "XX" y "NN" respectivamente). Las
estaciones del programa de monitoreo "Laguna Garzón" presentan esta categoría
en la columna "departamento" (Se encuentran entre Maldonado y Rocha).
}
\details{
Fecha de extraccion: 2021-09-17
}
\keyword{datasets}
