% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_estacion}
\alias{sia_estacion}
\title{Estaciones de monitoreo}
\format{
Tabla con 402 filas y 17
columnas:

\describe{

\item{codigo_pto}{Código de letras y números para cada estación}

\item{estacion}{Nombre de la estación (largo; puede ser una descripción de
cómo llegar, etc...)}

\item{latitud}{}

\item{longitud}{}

\item{gid}{Identificador único (número entero) para el contexto de trabajo
con GIS}

\item{id_playa}{..completar..}

\item{prog_monitoreo}{Identificador único (id) del programa de monitoreo
correspondiente a cada estación.}

\item{id}{Identificador único (número entero) de la estación.}

\item{version}{..completar..}

\item{tipo_punto_id}{Número que indica el id del tipo de muestras tomadas:
superficie o fondo (1 o 2 respectivamente)}

\item{estacion_asociada_id}{..completar..}

\item{departamento}{Número que indica el id del departamento en el que se
encuentra la estación}

\item{sub_cuenca}{Número que indica el id de la subcuenca en el que se
encuentra la estación}

\item{orden_ingreso}{..completar..}

\item{ingreso_interno}{..completar..}

\item{matriz_estacion}{Número que indica el id de la matriz asociada a la
estación (de momento: 6 = Aguas superficiales y 11 = Sedimentos;
15/10/2020)}

\item{estacion_activa}{..completar..}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_estacion
}
\description{
Es la lista completa de estaciones de monitoreo para las matrices Aguas
superficiales y Sedimentos. Incluye las columnas \code{prog_monitoreo} y
\code{matriz_estacion}, las cuales son la razón por las cuales las estaciones
son usadas como referencia para adivinar si un conjunto de datos pertenece a
una determinada matriz y programa. También se usa para determinar la matriz
de cada programa de monitoreo.
}
\details{
Fecha de extraccion: 2021-08-05
}
\keyword{datasets}
