% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_institucion}
\alias{sia_institucion}
\title{Lista de instituciones}
\format{
Tabla con 27 filas y 2
columnas:

\describe{

\item{id_institucion}{Número único (natural) para cada institución.}

\item{nombre}{Nombre de la institución (texto).}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_institucion
}
\description{
Instituciones asociadas a los usuarios que suben los datos al SIA.
}
\details{
Fecha de extraccion: 2021-09-17
}
\keyword{datasets}
