% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_matriz}
\alias{sia_matriz}
\title{Lista de matrices ambientales}
\format{
Tabla de 15 filas y 5 columnas:

\describe {

\item{id_matriz}{}

\item{nombre}{Nombre de la matriz}

\item{codigo}{Código o nombre corto de la matriz}

\item{vigente}{..completar..}

\item{fca_habilitado}{..completar..}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_matriz
}
\description{
La tabla tiene los nombres y abreviaciones de las matrices ambientales
definidas e infambientalbd, así como el id único de cada clase.
}
\details{
Fecha de extraccion: 2020-10-22
}
\keyword{datasets}
