% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_muestra}
\alias{sia_muestra}
\title{Inventario de muestras de infambientalbd}
\format{
Tabla con 19952 filas y 11
columnas:

\describe{

\item{id_muestra}{integer. Número único identificador de la muestra}

\item{id_institución}{integer. Identificador de la institución a la que
pertenece el usuario que cargó los datos en el SIA}

\item{periodo}{character. Texto en formato MES AÑO (ej.: JUNIO 2012). No
necesariamente se corresponde con \code{fecha_muestra}, ya que se trata de un
campo de identificación de la campaña de muestreo}

\item{fecha_muestra}{date. Fecha en que se tomó la muestra.}

\item{hora_muestra}{character. Hora en que se tomó la muestra (en el
campo).}

\item{usuario}{character. Usuario que ingresó los datos al SIA (no
necesariamente corresponde con quien/es lo colectaron en campo)}

\item{fecha_ingreso}{datetime. Fecha y hora en que se ingresó la muestra al
SIA.}

\item{observaciones}{character. Observaciones.}

\item{id_estacion}{integer. Identificador de la estación de monitoreo.}

\item{replica}{logical. ..completar..}

\item{nro_muestra}{integer. Número de muestra asignado por SILAD. No es
único para esta tabla (no confundir con id_muestra).}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_muestra
}
\description{
Esta tabla guarda los metadatos generales de cada muestra. Una muestra se
corresponde con el muestreo en campo en un sitio en particular (estación) y
una fecha determinada.
}
\details{
Fecha de extraccion: 2021-08-05
}
\keyword{datasets}
