% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_param_unidad}
\alias{sia_param_unidad}
\title{Lista de correspondencias entre parámetros y unidades}
\format{
Tabla con 1321 filas y
4)` columnas:

\describe{

\item{id}{identificador único para las entradas de esta tabla (i.e.:
combinaciones de \code{id_unidad_medida}, \code{id_parametro} e \code{id_matriz})}

\item{id_unidad_medida}{Número que indica el id de la unidad de medida
correspondiente}

\item{id_parametro}{Número que indica el id del parámetro correspondiente}

\item{id_matriz}{Número que indica el id de la matriz ambiental
correspondiente}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_param_unidad
}
\description{
Tabla que vincula cada parámetro con su unidad correspondiente, según la
matriz ambiental. Necesaria para conectar \code{\link{sia_parametro}} con
\code{\link{sia_unidad}}
}
\details{
Fecha de extraccion: 2020-10-22
}
\keyword{datasets}
