% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_parametro}
\alias{sia_parametro}
\title{Lista de parámetros de SIA}
\format{
Tabla con 404 filas y 7
columnas:

\describe{

\item{id_parametro}{id: número natural único para cada fila}

\item{parametro}{Nombre (largo) del parámetro. Puede ser una frase corta.
Texto.}

\item{enumerado}{..completar..}

\item{nombre_clave}{Nombre corto o código creado para cada parámetro
(debería ser único). Texto.}

\item{decimales}{..completar..}

\item{par_vigente}{..completar..}

\item{codigo_airviro}{..completar..}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_parametro
}
\description{
Tabla con todo los parámetros contemplados por la base de datos
infambientalbd. Para conectarla con \code{\link{sia_unidad}}, es necesaria la
tabla \code{\link{sia_param_unidad}}.
}
\details{
Fecha de extraccion: 2021-09-17
}
\keyword{datasets}
