% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_programa}
\alias{sia_programa}
\title{Lista de programas de monitoreo}
\format{
Tabla con 34 filas y 6
columnas:

\describe{

\item{id_programa}{\code{integer}. Número único para cada programa.}

\item{nombre_programa}{Nombre (largo) del programa de monitoreo}

\item{codigo_programa}{Código (corto) que identifica a cada programa}

\item{visible_externos}{..completar..}

\item{version}{..completar..}

\item{id_programa_silad}{..completar..}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_programa
}
\description{
Los programas de monitoreo registrados en el SIA (base de datos
infambientalbd). Corresponden a las matrices ambientales "Aguas
superficiales" y "Sedimentos"
}
\details{
Fecha de extraccion: 2021-09-17
}
\keyword{datasets}
