% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_sub_cuenca}
\alias{sia_sub_cuenca}
\title{Lista de subcuencas}
\format{
Tabla con 302 filas y
7)` columnas:

\describe{

\item{id}{integer. Número asignado a la subcuenca}

\item{sub_cue_nombre}{Nombre de la subcuenca}

\item{sub_cue_cuenca_id}{integer. Id de la cuenca asociada, correspondiente
a la tabla \code{\link{sia_cuenca}}}

\item{sub_cue_curso}{Columna asociada a otras tablas de
\code{ifnambientalbd}}

\item{sub_cue_cod_sc}{Columna asociada a otras tablas de
\code{ifnambientalbd}}

\item{cursos_de_agua}{Columna asociada a otras tablas de
\code{ifnambientalbd}}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_sub_cuenca
}
\description{
La tabla de subcuencas tiene 302 cuencas definidas,
asociadas a las cuencas de \code{\link{sia_cuenca}} a través de la columna
\code{sub_cue_cuenca_id}.
}
\details{
Las columnas \code{sub_cue_curso}, \code{sub_cue_cod_sc} y
\code{cursos_de_agua} refieren a otras tablas de \code{infambientalbd} no
incluidas en el presente paquete.

Fecha de extraccion: 2021-08-05
}
\examples{
dplyr::left_join(dplyr::select(sia_sub_cuenca, id:sub_cue_cuenca_id),
                 sia_cuenca, by = c('sub_cue_cuenca_id' = 'id'))

}
\keyword{datasets}
