% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_tipo_punto_estacion}
\alias{sia_tipo_punto_estacion}
\title{Tipo de estacion}
\format{
Tabla con 2 filas y
3)` columnas:

\describe{

\item{id}{Identificador único para las entradas de esta tabla}

\item{tip_pun_est_descripcion}{Descripción del tipo de punto ("SUPERFICIE"
o "FONDO")}

\item{tip_pun_est_codigo}{Identificador único (númerico) del tipo de punto:
superficie o fondo (1 o 2 respectivamente)}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_tipo_punto_estacion
}
\description{
Tipos de estaciones de monitoreo, según (la profundidad de) las muestras que
se toman
}
\details{
Fecha de extraccion: 2021-08-05
}
\keyword{datasets}
