% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_unidad}
\alias{sia_unidad}
\title{Lista de las unidades de medidas contempladas por el SIA}
\format{
Tabla con 118 filas y 3 columnas:

\describe{

\item{id}{integer. Identificador único de cada unidad de medida.}

\item{uni_nombre}{Texto para cada unidad de medida (ej: "mg CaCO3/L")}

\item{uni_factor_conversion}{..completar..}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_unidad
}
\description{
La tabla oficial de unidades de medida registradas en el SIA. Para vincular
las unidades con \code{\link{sia_parametro}} es necesaria la tabla
\code{\link{sia_param_unidad}}
}
\details{
Fecha de extraccion: 2021-09-17
}
\keyword{datasets}
