% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siabox.extraction.meta}
\alias{siabox.extraction.meta}
\title{Metadatos de las tablas importadas del SIA}
\format{
An object of class \code{character} of length 9.
}
\usage{
siabox.extraction.meta
}
\description{
Combina el output de \code{Sys.info()} con \code{Sys.Date()} para guardar la
información relativa a la máquina y fecha en que se hicieron las extracciones
de las tablas provenientes del SIA. La fecha en prticular se usa luego para
actualizar la documentación.
}
\examples{
## Fecha de extraccion:
siabox.extraction.meta['extraction.date']
}
\keyword{datasets}
