% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{t_eti_add}
\alias{t_eti_add}
\title{Crea o modifica tabla de etiquetas}
\usage{
t_eti_add(id_parametro = NULL, etiqueta)
}
\arguments{
\item{id_parametro}{Vector integer}

\item{etiqueta}{Vector character o expression}
}
\value{

}
\description{
Sin argumentos definidos, genera una tabla de etiquetas estándar, basada en
las tablas del SIA (infambientalbd): \code{sia_parametro},
\code{sia_param_unidad} y \code{sia_unidad}. Con argumentos, agrega o
modifica las etiquetas de uno o más parámetros.
}
\details{
La idea de la tabla de etiquetas es generar una referencia que se usará en
funciones para crear gráficos, tablas, etc...
}
\examples{
# Sin ninguna etiqueta nueva:
t_eti <- t_eti_add()
plot(0, 0, xlab = dplyr::filter(t_eti, id_parametro == 2005)[[2]])

# Modificando un parámetro existente:
t_eti <- t_eti_add(2005L, expression('AlcT (mg CaCO' [3] * ' /L)'))
plot(0, 0, xlab = dplyr::filter(t_eti, id_parametro == 2005)[[2]])

# Modificando un parámetro existente, agregando uno nuevo y mezclando
# etiquetas del tipo expression y character:
t_eti <- t_eti_add(
  c(2005L, 9999L),
  c(expression('AlcT (mg CaCO' [3] * ' /L)'), "YOLO (Mb/min)")
)
plot(0, 0,
     xlab = dplyr::filter(t_eti, id_parametro == 2005L)[[2]],
     ylab = dplyr::filter(t_eti, id_parametro == 9999L)[[2]])
}
\seealso{
\code{\link{t_eti_base}}
}
