% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t_eti_base}
\alias{t_eti_base}
\title{Tabla de etiquetas base}
\format{
Tabla con 302 filas y 2 columnas:

\describe{

\item{id_parametro}{integer. Identificador único de cada parámetro.}

\item{etiqueta}{character. Texto de etiqueta para gráficos, etc...}

}
}
\source{
\code{infambientalbd}
}
\usage{
t_eti_base
}
\description{
Creada con \code{\link{make_t_eti_base()}}
}
\details{
Fecha de creación: 2020-10-29
}
\keyword{datasets}
