% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tipos_de_dato}
\alias{tipos_de_dato}
\title{Clasificación de los datos según el valore escrito}
\format{
Tabla con 7 filas y 2 columnas:

\describe{

\item{tipo_dato}{character. Nombre de las categorías (tipos) de datos.}

\item{id_tipo_dato}{integer. Id, número único entero que identifica a cada
tipo de dato}

}
}
\usage{
tipos_de_dato
}
\description{
Esta tabla registra los tipos de dato encontrados en infambientalbd
(específicamente, en la columna \code{valor_minimo_str} de la tabla
\code{\link{sia_datos_muestra_parametros}}).
}
\details{
Se espera que los tipos de dato (columna \code{tipo_dato}) se autoexpliquen. En
particular, las categorías \verb{<X} y \verb{>X} refieren a casos en los que se X es un
valor numérico, como "<2.0" o ">5000".
}
\examples{
# Para crear la tabla:
tipos_de_dato <- tibble::tibble(
  tipo_dato = c("NUMERICO", "<LD", "<LC", "LD<X<LC", "<X", ">X", "OTRO"),
  id_tipo_dato = 1:7
)
}
\seealso{
\code{\link{clasif_tipo_dato}}
}
\keyword{datasets}
