% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{tsummary}
\alias{tsummary}
\title{Summary de una columna en formato tabla}
\usage{
tsummary(.data, columna)
}
\arguments{
\item{.data}{Tabla (\code{\link[base]{data.frame}}) con datos.}

\item{columna}{Columna numerica de \code{.data}, escrita sin comillas.}
}
\value{
Devuelve un summary de los valores de la columna en cuestion,
ordenado de forma tal que cada estadistico es una columna. Los estadisticos
tomados son: n (numero de valores), Min (minimo), '1er Cu' (primer
cuartil), Media (promedio), Mediana (i.e.: segundo cuartil), '3er Cu'
(tercer cuartil) y Max (maximo).

En caso de que \code{.data} sea de clase \code{\link[dplyr]{grouped_df}},
devuelve una fila por cada grupo de \code{.data} con los estadisticos
correspondientes.
}
\description{
Función alternativa a \code{\link[base]{summary}}, pensada para trabajar
fluidamente con tablas de datos y el tidyverse.
}
\examples{
tsummary(tibble::tibble(x = c(3, 3, 5, 2.2, 1, 3.4, 7.6)), x)
datos_sia \%>\% tsummary(valor)
datos_sia \%>\% dplyr::group_by(id_parametro) \%>\% tsummary(valor)
# Puede demorar unos segundos:
datos_sia \%>\%
  dplyr::filter(id_programa \%in\% c(4L, 7L)) \%>\%
  dplyr::group_by(id_parametro, id_programa, id_estacion) \%>\%
  tsummary(valor) \%>\%
  dplyr::left_join(sia_estacion[c(1, 8)], by = c("id_estacion" = "id")) \%>\%
  dplyr::left_join(sia_parametro)
}
