% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles.R
\name{unipar}
\alias{unipar}
\title{Ver unidades de un parámetro SIA}
\usage{
unipar(id_parametro, id_matriz = 6L, nombre_clave)
}
\arguments{
\item{id_parametro}{integer. id del parámetro (ver
\code{\link{sia_parametro}})}

\item{id_matriz}{integer. id de la matriz (ver \code{\link{sia_matriz}})}

\item{nombre_clave}{character. Código o nombre clave del parámetro (ver
\code{\link{sia_parametro}})}
}
\value{

}
\description{
Ver unidades de un parámetro SIA
}
\examples{
unipar(c(2098, 2005))
unipar(nombre_clave = "^PT$")
unipar(nombre_clave = "Colif")
}
\seealso{
\code{\link{sia_parametro}}
}
