% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usuarios}
\alias{usuarios}
\title{Tabla con lista de usuarios}
\format{
Tabla con 38 filas y
1 columnas:

\describe{

\item{usuario}{character. Nombre de usuario}

}
}
\usage{
usuarios
}
\description{
La lista de usuarios que han cargado datos de muestras a las tablas del SIA.
}
\seealso{
\code{\link{sia_datos_muestra_parametros}}
}
\keyword{datasets}
