% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacion.R
\name{v_valor}
\alias{v_valor}
\title{Evaluación de valores}
\usage{
v_valor(.data)
}
\arguments{
\item{.data}{}
}
\value{
Tabla con varias columnas booleanas que indican resultados de varias
evaluaciones:
\itemize{
\item Si hay valores menores al mínimo del rango normal esperado.
\item Si hay valores mayores al máximo del rango normal esperado.
\item Si hay valores menores al límite de detección.
\item Si hay valores menores al límite de cuantificación.
\item Si hay límites de detecciones mayores a los límites de cuantificaciones.
}
}
\description{
Evaluación de valores
}
\examples{
alt <- siabox::datos_sia \%>\%
  dplyr::mutate(
    limite_deteccion = as.numeric(limpia_num(limite_deteccion)),
    limite_cuantificacion = as.numeric(limpia_num(limite_cuantificacion))
  ) \%>\%
  dplyr::left_join(
    dplyr::select(siabox::sitio,
                  id_sitio, id_matriz, id_estacion = id_interno),
    by = c("id_matriz", "id_estacion")
  ) \%>\%
  dplyr::select(id_sitio, codigo_pto, fecha_muestra, id_matriz, id_parametro,
                param, valor, limite_deteccion, limite_cuantificacion) \%>\%
  v_valor \%>\%
  dplyr::filter(menor_minimo | mayor_maximo | valor_menor_ld |
                  valor_menor_lc | ld_mayor_lc) \%>\%
  dplyr::filter(mayor_maximo) \%>\%
  dplyr::select(codigo_pto, param, fecha_muestra,
                valor_max, valor_max_sitio, vsup, valor)
}
