% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracciones_infambiental.R
\name{valores_numericos}
\alias{valores_numericos}
\title{Convertir valores del SIA en numéricos}
\usage{
valores_numericos(.data, filtrar_no_num = FALSE, metodo = "simple")
}
\arguments{
\item{.data}{\code{data.frame} con datos provenientes de la base de datos del SIA
(infambientalbd), con al menos tres columnas: \code{valor_minimo_str},
\code{limite_deteccion} y \code{limite_cuantificacion} (ver detalles).}

\item{filtrar_no_num}{\code{logical} (bandera). ¿Conservar los valores que no se
pudieron convertir en numéricos?}

\item{metodo}{character. Opciones: "sin_cambios", "basico", "simple",
"informe". Ver \code{\link{clasif_tipo_dato}}}
}
\value{
\code{tibble} con datos originales y una columna numérica extra, \code{valor},
cuyos valores son el resultado de sustitución realizadas con expresiones
regulares. Ver detalles.
}
\description{
Agrega una columna,llamada \code{valor}, de clase numeric, a una tabla con
datos del SIA,  con los valores originales (\code{valor_minimo_str}) convertidos a
numéricos. Los requisitos se exponen en detalles.
}
\details{
Esta función se creó en el contexto de analizar datos numéricos para
validación, pero puede usarse potencialmente para otras tareas, tales como
análisis y visualización de datos provenientes del SIA.

Requiere que .data incluya columnas con valores y límites (detección y
cuantificación), con los mismos nombres que usa la tabla
\code{\link{sia_datos_muestra_parametros}} de la base de datos
infambientalbd (SIA).:
\itemize{
\item \code{valor_minimo_str} (character)
\item \code{limite_deteccion} (character)
\item \code{limite_cuantificacion} (character)
}

Nota: esto implica que .data tiene formato "largo" (ver
\code{\link[tidyr]{pivot_longer}}), es decir, que en lugar de una columna
para cada parámetro, se incluye una (o más) columna con el nombre del
parámetro correspondiente a cada fila.

En la columna \code{valor} de la salida, se encuentran los valores de los
parámetros, convertidos en numéricos. Sin importar el método elegido, la
modificación mínima, además de aplicar \code{\link[base]{as.numeric}}, es
cambiar comas, comas repetidas y puntos repetidos por un único punto
(marcador de decimales).

Los métodos contemplados implican las siguientes conversiones (X representa
un valor numérico):

\describe{

\item{sin_cambios}{No se aplican cambios, de forma que la columna \code{valor}
resultante es character e idéntica a \code{valor_minimo_str}}

\item{basico}{Se aplica \code{\link[base]{as.numeric}} a la columna
\code{valor_minimo_str}, luego de hacer unos cambios mínimos (corrigiendo comas
por puntos, etc...). El resultado concreto es que todo lo que es
reconocible como valor numérico, se mantiene en \code{valor}, mientras que el
resto serán \code{NA}s}

\item{simple}{

\itemize{

\item <LD = LD

\item <LC = LC

\item <X = X

\item >X = X

}

}

\item{informe}{

\itemize{

\item <LD = LD

\item <LC = LC/2

\item LD<X<LC = (LD + LC) / 2

\item <X = X

\item >X = X

}

}

}
}
\examples{
d <- datos_sia \%>\%
  dplyr::select(id_parametro, valor_minimo_str,
                limite_deteccion, limite_cuantificacion)

valores_numericos(d)
valores_numericos(d, metodo = "informe")
valores_numericos(d, metodo = "simple")
# Porcentajes de algunos tipos de dato:
valores_numericos(d, metodo = "informe") \%>\%
  dplyr::group_by(id_parametro) \%>\%
  dplyr::summarise(
    porcentaje_numerico = sum(id_tipo_dato == 1L) / dplyr::n(),
    porcentaje_menor_lim = sum(id_tipo_dato \%in\% 2:4) / dplyr::n()
    )

filtrar_datos(datos_sia, id_programa = 3) \%>\%
  valores_numericos(metodo = "basico")
}
