% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracciones_infambiental.R
\name{ancho}
\alias{ancho}
\title{Formato ancho}
\usage{
ancho(.data, unidades = FALSE)
}
\arguments{
\item{.data}{Tabla de datos obtenida con \code{\link{consulta_muestras}},
posiblemente modificada con valores_numericos y con columnas agregadas por
\code{\link[dplyr]{left_join}}.}

\item{unidaes}{TRUE o FALSE. Determina si se agregan las unidades a las columnas de los parámetros (ej.: 'SatO (\%)' en lugar de 'SatO')}
}
\value{

}
\description{
Ensanchar datos provenientes de consulta_muestras. Espera la presencia de
ciertas columnas y una sóla matriz ambiental. Ver detalles.
}
\details{
Espera que existan las columnas 'valor' y 'param'. En caso de no
encontrarlas las creará. La primera contiene los valores de los parámetros
muestreados, ya sea en formato numérico o en texto. La segunda debería
contener el código del parámetro junto a las unidades de medida entre
paréntesis (ej.: \code{'OD (mg/L)'}). En caso de haber una columna llamada
'codigo_nuevo', será utilizada para crear 'param', de lo contrario usará la
columna 'nombre_clave' (de la tabla parametro, de infambientalbd).

Para facilitar la compatibilidad, esta función además verifica la presencia
de otras columnas agregadas en el código de la app iSIA (en el reactive
\code{datos_extraccion}, del server.R): codigo_nuevo, parametro, grupo,
id_tipo_dato y tipo_dato.
}
\examples{
d <- filtrar_datos(datos_sia,
                   id_programa = 10L,
                   rango_fechas = c("2019-01-01", "2020-12-31"),
                   id_parametro = c(2099, 2098)) \%>\%
  dplyr::select(codigo_pto, fecha_muestra, id_parametro, param, id_unidad,
                uni_nombre, limite_deteccion, limite_cuantificacion, valor)
print(d)
ancho(d)

# Cómo usar ancho en lugar de ancho_old:
datos_sia \%>\%
  # Primero filtrar para tener sólo 2 parámetros:
  dplyr::filter(id_programa == 4, id_parametro \%in\% c(2017, 2021)) \%>\%
  ancho \%>\%
  dplyr::select(SatO, OD) \%>\%
  plot
}
