% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos_sia}
\alias{datos_sia}
\alias{datos_sia,}
\alias{datos_sia_sed}
\title{Datos extraídos del SIA.}
\format{
Tabla con 258379 filas y 39 columnas:

\describe{

\item{id_muestra}{Número único que identifica cada muestra ingresada al
banco de datos del SIA.}

\item{nro_muestra}{Número único que identifica cada muestra ingresada al
banco de datos del SILAD.}

\item{nombre_programa}{Nombre del programa de monitoreo al que corresponde
la muestra.}

\item{id_programa}{Número único que identifica al programa de monitoreo.}

\item{cue_nombre}{Nombre de la cuenca en la que se encuentra la estación en
donde se tomó la muestra.}

\item{id_cuenca}{Número único que identifica la cuenca en la que se
encuentra la estación en donde se tomó la muestra.}

\item{sub_cue_nombre}{Nombre de la sub-cuenca en la que se encuentra la
estación en donde se tomó la muestra.}

\item{id_sub_cuenca}{Número único que identifica la sub-cuenca en la que se
encuentra la estación en donde se tomó la muestra.}

\item{codigo_pto}{Código que identifica la estación o punto de monitoreo,
lugar en donde se tomó la muestra.}

\item{id_estacion}{Número único que identifica a cada estación.}

\item{tipo_punto_id}{Número único que identifica a cada tipo de punto.}

\item{tip_pun_est_descripcion}{Texto descriptivo de cada tipo de punto.}

\item{id_depto}{Número único que identifica a cada departamento.}

\item{departamento}{Nombre de cada departamento.}

\item{id_institucion}{Número único que identifica a cada institución.}

\item{institucion}{Nombre e cada institución.}

\item{usuario}{Nombre de usuario que ingresó la muestra a la base de
datos.}

\item{periodo}{Título que identifica a cada campaña de muestreo; se compone
del nombre del mes y el año (ej.: JUNIO 2019).}

\item{anio}{Año en que fue tomada la muestra.}

\item{mes}{Mes en que fue tomada la muestra.}

\item{anio_mes}{Año y mes en que fue tomada la muestra.}

\item{fecha_muestra}{Fecha del día en que fue tomada la muestra.}

\item{fecha_hora}{Hora en que fue tomada la muestra.}

\item{observaciones}{Observaciones correspondientes a cada muestra o
valor.}

\item{id_matriz}{Número único que identifica a cada matriz ambiental.}

\item{id_parametro}{Número único que identifica a cada parámetro.}

\item{parametro}{Descripción extendida de cada parámetro.}

\item{nombre_clave}{Código textual único que identifica a cada parámetro.}

\item{id_unidad}{Número único que identifica a cada unidad de medida.}

\item{uni_nombre}{Texto que describe las unidades de medida (ej.: \verb{mg NO2-N/L}).}

\item{valor_minimo_str}{Texto con el valor ingresado para el parámetro en
cada muestra.}

\item{limite_deteccion}{Texto con el valor ingresado para el límite de
detección del parámetro en cada muestra.}

\item{limite_cuantificacion}{Texto con el valor ingresado para el límite de
cuantificación del parámetro en cada muestra.}

\item{valor}{numeric. Valor ingresado para cada parámetro, obtenido a
partir del uso de la función \code{link{valores_numericos}} con la opción
\code{metodo = "informe"}.}

\item{id_tipo_dato}{Es un número asignado a cada valor. Se trata de una
clasificación de cada dato según siete categorías (ver
\code{\link{tipos_de_dato}})}

\item{nombre_subcuenca_informes}{Nombre dado a cada subcuenca dentro de los
informes anuales de los programas de monitoreo.}

\item{codigo_pto_mnod}{Nombre de la estación modificado, siguiendo los
códigos presentes en la tabla \code{\link{cuencas_informes}}}

\item{grupo}{Grupo al que pertenece el parámetro (ej.: Parámetros de
Biológicos, Parámetros de Ecotoxicidad). No todos los parámetros tienen un
grupo asignado.}

\item{param}{Nombre corto del parámetro (mejorado a partir de
\code{\link{codigos_param}}; en caso de parámetros que no figuran en esa
tabla, se usa el \code{nombre_clave} de \code{\link{sia_parametro}})}

}
}
\source{
\code{infambientalbd}
}
\usage{
datos_sia
}
\description{
La tibble \code{datos_sia} tiene datos de matriz Aguas superficiales y
\code{datos_sia_sed} de Sedimentos.
}
\details{
Fecha de extracción: `2021-08-05
}
\seealso{
\code{\link{tipos_de_dato}}, \code{\link{consulta_muestras}},
\code{\link{codigos_param}}
}
\keyword{datasets}
