% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{filtrar_datos}
\alias{filtrar_datos}
\title{Filtrar datos_sia}
\usage{
filtrar_datos(
  .data,
  id_programa,
  id_matriz = 6L,
  rango_fechas = NULL,
  id_parametro = NULL,
  id_estacion = NULL,
  orden_est = NULL,
  tipo_punto_id = 1L
)
}
\arguments{
\item{.data}{Tabla con datos extraidos del SIA (\code{\link{datos_sia}} en
principio)}

\item{id_programa}{integer. Un sólo valor que identifica al programa.}

\item{id_matriz}{integer. Un sólo valor que identifica a la matriz.}

\item{rango_fechas}{character. Vector de dos valores (fecha inicial y fecha
final), en formato \code{AAAA-MM-DD}.}

\item{id_parametro}{integer. Vector de valores que identifican los
parámetros.}

\item{id_estacion}{integer. Vector de valores que identifican las estaciones
(sitios) de monitoreo.}

\item{orden_est}{character. Vector con los nombres de las estaciones en el
orden deseado para las gráficas u otros usos.}

\item{tipo_punto_id}{integer. Vector que define los tipos de punto usados: 1
= \code{SUPERFICIE}, 2 = \code{FONDO}, 1:2 = Ambos tipos.}
}
\value{

}
\description{
Función que filtra la tabla \code{\link{datos_sia}} (u otra tabla con un
conjunto de columnas adecuado) de manera similar a la forma en que se extraen
los datos en la aplicación shiny iSIA.
}
\details{
El uso del argumento \code{orden_est} equivale, salvo algunas excepciones (que son
anunciadas con warnings), a ejecutar:

\code{.data$codigo_pto <- factor(.data$codigo_pto, levels = orden_est)}

Nota: \code{codigo_pto} es la columna con los nombres de las estaciones.
}
\examples{
filtrar_datos(datos_sia, 5, 6)
filtrar_datos(datos_sia, 5, 11)
}
