% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{g_long}
\alias{g_long}
\title{Graficar valores anuales longitudinales}
\usage{
g_long(
  .data,
  id_parametro,
  anio,
  colores_meses = (scales::hue_pal())(12),
  horiz
)
}
\arguments{
\item{.data}{data.frame. Datos del SIA (ver \code{\link{datos_sia}} y
ejemplos)}

\item{id_parametro}{integer. Id de uno/escalar (\code{g_mes_pto}) o varios
elementos (\code{g_mes_pto_all}). Se debe(n) corresponder con los ids
encontrados columna homónima de \code{\link{sia_parametro}}}

\item{anio}{intger. Año}

\item{colores_meses}{character. Paleta de colores para usar en los distintos
meses.}

\item{horiz}{numeric. Vector con dos valores: min y max, para ser graficados
con líneas horizontales.}

\item{tabla_horiz}{data.frame. Debe tener las columnas \code{id_parametro},
\code{valor}, y \code{extremo} tal como en la tabla \code{\link{decreto}}}

\item{path}{character. Ruta de directorio en donde guardar las imágenes
generadas por \code{g_long}}
}
\value{

}
\description{
Gráfica los valores de los parámetros por estaciones, ordenadas en orden de
la cuenca
}
\examples{
d <- readRDS("tmp/datos_cuareim.rds")
decreto <- readRDS("../data/decreto.rds")
d$mes <- as.integer(d$mes)
d$anio <- as.integer(d$anio)
g_long(d, 2032, 2019L, horiz = c(min = 5, max = 28))
}
