% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles.R
\name{limpia_num}
\alias{limpia_num}
\title{Limpiar texto con valores numéricos}
\usage{
limpia_num(x)
}
\arguments{
\item{x}{character. Valores numéricos expresados en varias formas posibles.}
}
\value{

}
\description{
Quita espacios en blanco, y sustituye comas, comas
repetidas y puntos repetidos por un único punto (indicador de decimales). Se
basa en expresiones regulares y el paquete \code{stringr}.
}
\examples{
siabox:::limpia_num("2.3")
siabox:::limpia_num("2..3")
siabox:::limpia_num("2,3")
siabox:::limpia_num("2,,,,3")
siabox:::limpia_num("   2,,,, 3   ")
}
