% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sia_programa_parametro}
\alias{sia_programa_parametro}
\title{Parámetros en cada programa}
\format{
Tabla con 1673 filas y
3)` columnas:

\describe{

\item{id}{Identificador único para las entradas de esta tabla}

\item{id_programa}{Identificador de programa}

\item{id_parametro}{Identificador de parametro}

}
}
\source{
\code{infambientalbd}
}
\usage{
sia_programa_parametro
}
\description{
Tabla que (intenta) registrar los parametros que son monitoreados en cada
programa.
}
\details{
Fecha de extraccion: 2021-09-17
}
\keyword{datasets}
