% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t_eti_base}
\alias{t_eti_base}
\title{Tabla de etiquetas base}
\format{
Tabla con 302 filas y 2 columnas:

\describe{

\item{id_parametro}{integer. Identificador único de cada parámetro.}

\item{etiqueta}{character. Texto de etiqueta para gráficos, etc...}

}
}
\source{
\code{infambientalbd}
}
\usage{
t_eti_base
}
\description{
Tabla de base para hacer etiquetas para gráficos u otros usos,
basándose en las tablas del SIA (infambientalbd).
}
\details{
Fecha de creación: 2020-10-29
}
\examples{
# Código de creación:
t_eti_base <-
  sia_parametro \%>\%
  dplyr::left_join(dplyr::select(codigos_param, id_parametro, codigo_nuevo),
                   by = "id_parametro") \%>\%
  dplyr::mutate(param = dplyr::if_else(is.na(codigo_nuevo),
                                       nombre_clave, codigo_nuevo)) \%>\%
  dplyr::left_join(sia_param_unidad, by = "id_parametro") \%>\%
  dplyr::filter(id_matriz == 6L) \%>\%
  dplyr::left_join(sia_unidad, by = c("id_unidad_medida" = "id")) \%>\%
  dplyr::transmute(id_parametro,
                   etiqueta = paste0(param, " (", uni_nombre, ")"))
}
\keyword{datasets}
