% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles.R
\name{toascii}
\alias{toascii}
\title{Convertir texto en ASCII}
\usage{
toascii(texto)
}
\arguments{
\item{texto}{Un vector atomico character.}
}
\value{
Un vector atomico character, en el que todos los tildes, acentos,
dieresis, etc, son sustituidos por su version ASCII.
}
\description{
Convertir texto en ASCII
}
\examples{
toascii(c("sabañón!", "ungüento", "nâo è graça"))
}
